+++ vscode/src/services/StatusBar.ts
@@ -132,9 +132,9 @@ export function createStatusBar(): CodyStatusBar {
                     true
                 ),
                 createFeatureToggle(
-                    'Symf Context',
-                    'Experimental',
-                    'Enable context fetched via symf',
+                    'Search Context',
+                    'Beta',
+                    'Enable using the natural language search index as an Enhanced Context chat source',
                     'cody.experimental.symfContext',
                     c => c.experimentalSymfContext,
                     false